pyproject.toml
[tool.poetry]
name = "llamarag"
version = "0.1.0"
description = ""
authors = ["nicolapace <nicola.pace97@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.3.9"
langchain-huggingface = "^0.1.2"
langchain-openai = "^0.2.10"
langchain-chroma = "^0.1.4"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
